styles: linter files

+++ contracts/PoolInterface.sol
@@ -2,7 +2,9 @@
 pragma solidity >=0.7.0 <=0.9.0;
 
 interface PoolInterface {
-    function approve() external  returns(bool);
-    function withDraw() external returns(bool);
-    function allowance() external view returns (uint);
+    function approve() external returns (bool);
+
+    function withDraw() external returns (bool);
+
+    function allowance() external view returns (uint256);
 }
